実装要件を満たすように条件式を追記

+++ app/controllers/items_controller.rb
@@ -40,7 +40,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    item.destroy if user_signed_in? && item.user_id == current_user.id
     redirect_to root_path
   end
